@@ -1,8 +1,11 @@
-async function purgeContainers() {
-  const shell = require('shelljs');
+import shell from 'shelljs';
 
+async function purgeContainers() {
   shell.exec(`scripts/purge_docker_containers.sh`);
   console.log('[*] purged containers');
 }
 
-purgeContainers();
+purgeContainers().catch(error => {
+  console.log(error);
+  process.exit(1);
+});
