@@ -1,11 +1,14 @@
+import shell from 'shelljs';
+
 /**
  * purge chain data
  */
 async function purgeChains() {
-  const shell = require('shelljs');
-
   // purge_chains.sh
   shell.exec(`scripts/purge_chains.sh`);
   console.log('[*] purged chains');
 }
-purgeChains();
+purgeChains().catch(error => {
+  console.error(error);
+  process.exit(1);
+});
