@@ -1,290 +1,279 @@
-export const ERC20ABI = '[\n' +
-  '\t{\n' +
-  '\t\t"inputs": [\n' +
-  '\t\t\t{\n' +
-  '\t\t\t\t"internalType": "string",\n' +
-  '\t\t\t\t"name": "name",\n' +
-  '\t\t\t\t"type": "string"\n' +
-  '\t\t\t},\n' +
-  '\t\t\t{\n' +
-  '\t\t\t\t"internalType": "string",\n' +
-  '\t\t\t\t"name": "symbol",\n' +
-  '\t\t\t\t"type": "string"\n' +
-  '\t\t\t}\n' +
-  '\t\t],\n' +
-  '\t\t"stateMutability": "nonpayable",\n' +
-  '\t\t"type": "constructor"\n' +
-  '\t},\n' +
-  '\t{\n' +
-  '\t\t"anonymous": false,\n' +
-  '\t\t"inputs": [\n' +
-  '\t\t\t{\n' +
-  '\t\t\t\t"indexed": true,\n' +
-  '\t\t\t\t"internalType": "address",\n' +
-  '\t\t\t\t"name": "owner",\n' +
-  '\t\t\t\t"type": "address"\n' +
-  '\t\t\t},\n' +
-  '\t\t\t{\n' +
-  '\t\t\t\t"indexed": true,\n' +
-  '\t\t\t\t"internalType": "address",\n' +
-  '\t\t\t\t"name": "spender",\n' +
-  '\t\t\t\t"type": "address"\n' +
-  '\t\t\t},\n' +
-  '\t\t\t{\n' +
-  '\t\t\t\t"indexed": false,\n' +
-  '\t\t\t\t"internalType": "uint256",\n' +
-  '\t\t\t\t"name": "value",\n' +
-  '\t\t\t\t"type": "uint256"\n' +
-  '\t\t\t}\n' +
-  '\t\t],\n' +
-  '\t\t"name": "Approval",\n' +
-  '\t\t"type": "event"\n' +
-  '\t},\n' +
-  '\t{\n' +
-  '\t\t"anonymous": false,\n' +
-  '\t\t"inputs": [\n' +
-  '\t\t\t{\n' +
-  '\t\t\t\t"indexed": true,\n' +
-  '\t\t\t\t"internalType": "address",\n' +
-  '\t\t\t\t"name": "from",\n' +
-  '\t\t\t\t"type": "address"\n' +
-  '\t\t\t},\n' +
-  '\t\t\t{\n' +
-  '\t\t\t\t"indexed": true,\n' +
-  '\t\t\t\t"internalType": "address",\n' +
-  '\t\t\t\t"name": "to",\n' +
-  '\t\t\t\t"type": "address"\n' +
-  '\t\t\t},\n' +
-  '\t\t\t{\n' +
-  '\t\t\t\t"indexed": false,\n' +
-  '\t\t\t\t"internalType": "uint256",\n' +
-  '\t\t\t\t"name": "value",\n' +
-  '\t\t\t\t"type": "uint256"\n' +
-  '\t\t\t}\n' +
-  '\t\t],\n' +
-  '\t\t"name": "Transfer",\n' +
-  '\t\t"type": "event"\n' +
-  '\t},\n' +
-  '\t{\n' +
-  '\t\t"inputs": [\n' +
-  '\t\t\t{\n' +
-  '\t\t\t\t"internalType": "address",\n' +
-  '\t\t\t\t"name": "owner",\n' +
-  '\t\t\t\t"type": "address"\n' +
-  '\t\t\t},\n' +
-  '\t\t\t{\n' +
-  '\t\t\t\t"internalType": "address",\n' +
-  '\t\t\t\t"name": "spender",\n' +
-  '\t\t\t\t"type": "address"\n' +
-  '\t\t\t}\n' +
-  '\t\t],\n' +
-  '\t\t"name": "allowance",\n' +
-  '\t\t"outputs": [\n' +
-  '\t\t\t{\n' +
-  '\t\t\t\t"internalType": "uint256",\n' +
-  '\t\t\t\t"name": "",\n' +
-  '\t\t\t\t"type": "uint256"\n' +
-  '\t\t\t}\n' +
-  '\t\t],\n' +
-  '\t\t"stateMutability": "view",\n' +
-  '\t\t"type": "function"\n' +
-  '\t},\n' +
-  '\t{\n' +
-  '\t\t"inputs": [\n' +
-  '\t\t\t{\n' +
-  '\t\t\t\t"internalType": "address",\n' +
-  '\t\t\t\t"name": "spender",\n' +
-  '\t\t\t\t"type": "address"\n' +
-  '\t\t\t},\n' +
-  '\t\t\t{\n' +
-  '\t\t\t\t"internalType": "uint256",\n' +
-  '\t\t\t\t"name": "amount",\n' +
-  '\t\t\t\t"type": "uint256"\n' +
-  '\t\t\t}\n' +
-  '\t\t],\n' +
-  '\t\t"name": "approve",\n' +
-  '\t\t"outputs": [\n' +
-  '\t\t\t{\n' +
-  '\t\t\t\t"internalType": "bool",\n' +
-  '\t\t\t\t"name": "",\n' +
-  '\t\t\t\t"type": "bool"\n' +
-  '\t\t\t}\n' +
-  '\t\t],\n' +
-  '\t\t"stateMutability": "nonpayable",\n' +
-  '\t\t"type": "function"\n' +
-  '\t},\n' +
-  '\t{\n' +
-  '\t\t"inputs": [\n' +
-  '\t\t\t{\n' +
-  '\t\t\t\t"internalType": "address",\n' +
-  '\t\t\t\t"name": "account",\n' +
-  '\t\t\t\t"type": "address"\n' +
-  '\t\t\t}\n' +
-  '\t\t],\n' +
-  '\t\t"name": "balanceOf",\n' +
-  '\t\t"outputs": [\n' +
-  '\t\t\t{\n' +
-  '\t\t\t\t"internalType": "uint256",\n' +
-  '\t\t\t\t"name": "",\n' +
-  '\t\t\t\t"type": "uint256"\n' +
-  '\t\t\t}\n' +
-  '\t\t],\n' +
-  '\t\t"stateMutability": "view",\n' +
-  '\t\t"type": "function"\n' +
-  '\t},\n' +
-  '\t{\n' +
-  '\t\t"inputs": [],\n' +
-  '\t\t"name": "decimals",\n' +
-  '\t\t"outputs": [\n' +
-  '\t\t\t{\n' +
-  '\t\t\t\t"internalType": "uint8",\n' +
-  '\t\t\t\t"name": "",\n' +
-  '\t\t\t\t"type": "uint8"\n' +
-  '\t\t\t}\n' +
-  '\t\t],\n' +
-  '\t\t"stateMutability": "view",\n' +
-  '\t\t"type": "function"\n' +
-  '\t},\n' +
-  '\t{\n' +
-  '\t\t"inputs": [\n' +
-  '\t\t\t{\n' +
-  '\t\t\t\t"internalType": "address",\n' +
-  '\t\t\t\t"name": "spender",\n' +
-  '\t\t\t\t"type": "address"\n' +
-  '\t\t\t},\n' +
-  '\t\t\t{\n' +
-  '\t\t\t\t"internalType": "uint256",\n' +
-  '\t\t\t\t"name": "subtractedValue",\n' +
-  '\t\t\t\t"type": "uint256"\n' +
-  '\t\t\t}\n' +
-  '\t\t],\n' +
-  '\t\t"name": "decreaseAllowance",\n' +
-  '\t\t"outputs": [\n' +
-  '\t\t\t{\n' +
-  '\t\t\t\t"internalType": "bool",\n' +
-  '\t\t\t\t"name": "",\n' +
-  '\t\t\t\t"type": "bool"\n' +
-  '\t\t\t}\n' +
-  '\t\t],\n' +
-  '\t\t"stateMutability": "nonpayable",\n' +
-  '\t\t"type": "function"\n' +
-  '\t},\n' +
-  '\t{\n' +
-  '\t\t"inputs": [\n' +
-  '\t\t\t{\n' +
-  '\t\t\t\t"internalType": "address",\n' +
-  '\t\t\t\t"name": "spender",\n' +
-  '\t\t\t\t"type": "address"\n' +
-  '\t\t\t},\n' +
-  '\t\t\t{\n' +
-  '\t\t\t\t"internalType": "uint256",\n' +
-  '\t\t\t\t"name": "addedValue",\n' +
-  '\t\t\t\t"type": "uint256"\n' +
-  '\t\t\t}\n' +
-  '\t\t],\n' +
-  '\t\t"name": "increaseAllowance",\n' +
-  '\t\t"outputs": [\n' +
-  '\t\t\t{\n' +
-  '\t\t\t\t"internalType": "bool",\n' +
-  '\t\t\t\t"name": "",\n' +
-  '\t\t\t\t"type": "bool"\n' +
-  '\t\t\t}\n' +
-  '\t\t],\n' +
-  '\t\t"stateMutability": "nonpayable",\n' +
-  '\t\t"type": "function"\n' +
-  '\t},\n' +
-  '\t{\n' +
-  '\t\t"inputs": [],\n' +
-  '\t\t"name": "name",\n' +
-  '\t\t"outputs": [\n' +
-  '\t\t\t{\n' +
-  '\t\t\t\t"internalType": "string",\n' +
-  '\t\t\t\t"name": "",\n' +
-  '\t\t\t\t"type": "string"\n' +
-  '\t\t\t}\n' +
-  '\t\t],\n' +
-  '\t\t"stateMutability": "view",\n' +
-  '\t\t"type": "function"\n' +
-  '\t},\n' +
-  '\t{\n' +
-  '\t\t"inputs": [],\n' +
-  '\t\t"name": "symbol",\n' +
-  '\t\t"outputs": [\n' +
-  '\t\t\t{\n' +
-  '\t\t\t\t"internalType": "string",\n' +
-  '\t\t\t\t"name": "",\n' +
-  '\t\t\t\t"type": "string"\n' +
-  '\t\t\t}\n' +
-  '\t\t],\n' +
-  '\t\t"stateMutability": "view",\n' +
-  '\t\t"type": "function"\n' +
-  '\t},\n' +
-  '\t{\n' +
-  '\t\t"inputs": [],\n' +
-  '\t\t"name": "totalSupply",\n' +
-  '\t\t"outputs": [\n' +
-  '\t\t\t{\n' +
-  '\t\t\t\t"internalType": "uint256",\n' +
-  '\t\t\t\t"name": "",\n' +
-  '\t\t\t\t"type": "uint256"\n' +
-  '\t\t\t}\n' +
-  '\t\t],\n' +
-  '\t\t"stateMutability": "view",\n' +
-  '\t\t"type": "function"\n' +
-  '\t},\n' +
-  '\t{\n' +
-  '\t\t"inputs": [\n' +
-  '\t\t\t{\n' +
-  '\t\t\t\t"internalType": "address",\n' +
-  '\t\t\t\t"name": "to",\n' +
-  '\t\t\t\t"type": "address"\n' +
-  '\t\t\t},\n' +
-  '\t\t\t{\n' +
-  '\t\t\t\t"internalType": "uint256",\n' +
-  '\t\t\t\t"name": "amount",\n' +
-  '\t\t\t\t"type": "uint256"\n' +
-  '\t\t\t}\n' +
-  '\t\t],\n' +
-  '\t\t"name": "transfer",\n' +
-  '\t\t"outputs": [\n' +
-  '\t\t\t{\n' +
-  '\t\t\t\t"internalType": "bool",\n' +
-  '\t\t\t\t"name": "",\n' +
-  '\t\t\t\t"type": "bool"\n' +
-  '\t\t\t}\n' +
-  '\t\t],\n' +
-  '\t\t"stateMutability": "nonpayable",\n' +
-  '\t\t"type": "function"\n' +
-  '\t},\n' +
-  '\t{\n' +
-  '\t\t"inputs": [\n' +
-  '\t\t\t{\n' +
-  '\t\t\t\t"internalType": "address",\n' +
-  '\t\t\t\t"name": "from",\n' +
-  '\t\t\t\t"type": "address"\n' +
-  '\t\t\t},\n' +
-  '\t\t\t{\n' +
-  '\t\t\t\t"internalType": "address",\n' +
-  '\t\t\t\t"name": "to",\n' +
-  '\t\t\t\t"type": "address"\n' +
-  '\t\t\t},\n' +
-  '\t\t\t{\n' +
-  '\t\t\t\t"internalType": "uint256",\n' +
-  '\t\t\t\t"name": "amount",\n' +
-  '\t\t\t\t"type": "uint256"\n' +
-  '\t\t\t}\n' +
-  '\t\t],\n' +
-  '\t\t"name": "transferFrom",\n' +
-  '\t\t"outputs": [\n' +
-  '\t\t\t{\n' +
-  '\t\t\t\t"internalType": "bool",\n' +
-  '\t\t\t\t"name": "",\n' +
-  '\t\t\t\t"type": "bool"\n' +
-  '\t\t\t}\n' +
-  '\t\t],\n' +
-  '\t\t"stateMutability": "nonpayable",\n' +
-  '\t\t"type": "function"\n' +
-  '\t}\n' +
-  ']'
+export const ERC20_ABI = [
+  {
+    'inputs': [],
+    'stateMutability': 'nonpayable',
+    'type': 'constructor'
+  },
+  {
+    'anonymous': false,
+    'inputs': [
+      {
+        'indexed': true,
+        'internalType': 'address',
+        'name': 'owner',
+        'type': 'address'
+      },
+      {
+        'indexed': true,
+        'internalType': 'address',
+        'name': 'spender',
+        'type': 'address'
+      },
+      {
+        'indexed': false,
+        'internalType': 'uint256',
+        'name': 'value',
+        'type': 'uint256'
+      }
+    ],
+    'name': 'Approval',
+    'type': 'event'
+  },
+  {
+    'anonymous': false,
+    'inputs': [
+      {
+        'indexed': true,
+        'internalType': 'address',
+        'name': 'from',
+        'type': 'address'
+      },
+      {
+        'indexed': true,
+        'internalType': 'address',
+        'name': 'to',
+        'type': 'address'
+      },
+      {
+        'indexed': false,
+        'internalType': 'uint256',
+        'name': 'value',
+        'type': 'uint256'
+      }
+    ],
+    'name': 'Transfer',
+    'type': 'event'
+  },
+  {
+    'inputs': [
+      {
+        'internalType': 'address',
+        'name': 'owner',
+        'type': 'address'
+      },
+      {
+        'internalType': 'address',
+        'name': 'spender',
+        'type': 'address'
+      }
+    ],
+    'name': 'allowance',
+    'outputs': [
+      {
+        'internalType': 'uint256',
+        'name': '',
+        'type': 'uint256'
+      }
+    ],
+    'stateMutability': 'view',
+    'type': 'function'
+  },
+  {
+    'inputs': [
+      {
+        'internalType': 'address',
+        'name': 'spender',
+        'type': 'address'
+      },
+      {
+        'internalType': 'uint256',
+        'name': 'amount',
+        'type': 'uint256'
+      }
+    ],
+    'name': 'approve',
+    'outputs': [
+      {
+        'internalType': 'bool',
+        'name': '',
+        'type': 'bool'
+      }
+    ],
+    'stateMutability': 'nonpayable',
+    'type': 'function'
+  },
+  {
+    'inputs': [
+      {
+        'internalType': 'address',
+        'name': 'account',
+        'type': 'address'
+      }
+    ],
+    'name': 'balanceOf',
+    'outputs': [
+      {
+        'internalType': 'uint256',
+        'name': '',
+        'type': 'uint256'
+      }
+    ],
+    'stateMutability': 'view',
+    'type': 'function'
+  },
+  {
+    'inputs': [],
+    'name': 'decimals',
+    'outputs': [
+      {
+        'internalType': 'uint8',
+        'name': '',
+        'type': 'uint8'
+      }
+    ],
+    'stateMutability': 'view',
+    'type': 'function'
+  },
+  {
+    'inputs': [
+      {
+        'internalType': 'address',
+        'name': 'spender',
+        'type': 'address'
+      },
+      {
+        'internalType': 'uint256',
+        'name': 'subtractedValue',
+        'type': 'uint256'
+      }
+    ],
+    'name': 'decreaseAllowance',
+    'outputs': [
+      {
+        'internalType': 'bool',
+        'name': '',
+        'type': 'bool'
+      }
+    ],
+    'stateMutability': 'nonpayable',
+    'type': 'function'
+  },
+  {
+    'inputs': [
+      {
+        'internalType': 'address',
+        'name': 'spender',
+        'type': 'address'
+      },
+      {
+        'internalType': 'uint256',
+        'name': 'addedValue',
+        'type': 'uint256'
+      }
+    ],
+    'name': 'increaseAllowance',
+    'outputs': [
+      {
+        'internalType': 'bool',
+        'name': '',
+        'type': 'bool'
+      }
+    ],
+    'stateMutability': 'nonpayable',
+    'type': 'function'
+  },
+  {
+    'inputs': [],
+    'name': 'name',
+    'outputs': [
+      {
+        'internalType': 'string',
+        'name': '',
+        'type': 'string'
+      }
+    ],
+    'stateMutability': 'view',
+    'type': 'function'
+  },
+  {
+    'inputs': [],
+    'name': 'symbol',
+    'outputs': [
+      {
+        'internalType': 'string',
+        'name': '',
+        'type': 'string'
+      }
+    ],
+    'stateMutability': 'view',
+    'type': 'function'
+  },
+  {
+    'inputs': [],
+    'name': 'totalSupply',
+    'outputs': [
+      {
+        'internalType': 'uint256',
+        'name': '',
+        'type': 'uint256'
+      }
+    ],
+    'stateMutability': 'view',
+    'type': 'function'
+  },
+  {
+    'inputs': [
+      {
+        'internalType': 'address',
+        'name': 'to',
+        'type': 'address'
+      },
+      {
+        'internalType': 'uint256',
+        'name': 'amount',
+        'type': 'uint256'
+      }
+    ],
+    'name': 'transfer',
+    'outputs': [
+      {
+        'internalType': 'bool',
+        'name': '',
+        'type': 'bool'
+      }
+    ],
+    'stateMutability': 'nonpayable',
+    'type': 'function'
+  },
+  {
+    'inputs': [
+      {
+        'internalType': 'address',
+        'name': 'from',
+        'type': 'address'
+      },
+      {
+        'internalType': 'address',
+        'name': 'to',
+        'type': 'address'
+      },
+      {
+        'internalType': 'uint256',
+        'name': 'amount',
+        'type': 'uint256'
+      }
+    ],
+    'name': 'transferFrom',
+    'outputs': [
+      {
+        'internalType': 'bool',
+        'name': '',
+        'type': 'bool'
+      }
+    ],
+    'stateMutability': 'nonpayable',
+    'type': 'function'
+  }
+];
 
-export const ERC20BYTECODE = '60806040523480156200001157600080fd5b5060405162001b2b38038062001b2b833981810160405281019062000037919062000366565b818181600390805190602001906200005192919062000238565b5080600490805190602001906200006a92919062000238565b505050620000af3362000082620000b760201b60201c565b60ff16600a6200009391906200058a565b62989680620000a39190620006c7565b620000c060201b60201c565b5050620008bc565b60006012905090565b600073ffffffffffffffffffffffffffffffffffffffff168273ffffffffffffffffffffffffffffffffffffffff16141562000133576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016200012a9062000423565b60405180910390fd5b62000147600083836200022e60201b60201c565b80600260008282546200015b9190620004d2565b92505081905550806000808473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020600082825401925050819055508173ffffffffffffffffffffffffffffffffffffffff16600073ffffffffffffffffffffffffffffffffffffffff167fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef836040516200020e919062000445565b60405180910390a36200022a600083836200023360201b60201c565b5050565b505050565b505050565b828054620002469062000768565b90600052602060002090601f0160209004810192826200026a5760008555620002b6565b82601f106200028557805160ff1916838001178555620002b6565b82800160010185558215620002b6579182015b82811115620002b557825182559160200191906001019062000298565b5b509050620002c59190620002c9565b5090565b5b80821115620002e4576000816000905550600101620002ca565b5090565b6000620002ff620002f9846200048b565b62000462565b9050828152602081018484840111156200031e576200031d62000866565b5b6200032b84828562000732565b509392505050565b600082601f8301126200034b576200034a62000861565b5b81516200035d848260208601620002e8565b91505092915050565b6000806040838503121562000380576200037f62000870565b5b600083015167ffffffffffffffff811115620003a157620003a06200086b565b5b620003af8582860162000333565b925050602083015167ffffffffffffffff811115620003d357620003d26200086b565b5b620003e18582860162000333565b9150509250929050565b6000620003fa601f83620004c1565b9150620004078262000893565b602082019050919050565b6200041d8162000728565b82525050565b600060208201905081810360008301526200043e81620003eb565b9050919050565b60006020820190506200045c600083018462000412565b92915050565b60006200046e62000481565b90506200047c82826200079e565b919050565b6000604051905090565b600067ffffffffffffffff821115620004a957620004a862000832565b5b620004b48262000875565b9050602081019050919050565b600082825260208201905092915050565b6000620004df8262000728565b9150620004ec8362000728565b9250827fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff03821115620005245762000523620007d4565b5b828201905092915050565b6000808291508390505b60018511156200058157808604811115620005595762000558620007d4565b5b6001851615620005695780820291505b8081029050620005798562000886565b945062000539565b94509492505050565b6000620005978262000728565b9150620005a48362000728565b9250620005d37fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff8484620005db565b905092915050565b600082620005ed5760019050620006c0565b81620005fd5760009050620006c0565b8160018114620006165760028114620006215762000657565b6001915050620006c0565b60ff841115620006365762000635620007d4565b5b8360020a91508482111562000650576200064f620007d4565b5b50620006c0565b5060208310610133831016604e8410600b8410161715620006915782820a9050838111156200068b576200068a620007d4565b5b620006c0565b620006a084848460016200052f565b92509050818404811115620006ba57620006b9620007d4565b5b81810290505b9392505050565b6000620006d48262000728565b9150620006e18362000728565b9250817fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff04831182151516156200071d576200071c620007d4565b5b828202905092915050565b6000819050919050565b60005b838110156200075257808201518184015260208101905062000735565b8381111562000762576000848401525b50505050565b600060028204905060018216806200078157607f821691505b6020821081141562000798576200079762000803565b5b50919050565b620007a98262000875565b810181811067ffffffffffffffff82111715620007cb57620007ca62000832565b5b80604052505050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052601160045260246000fd5b7f4e487b7100000000000000000000000000000000000000000000000000000000600052602260045260246000fd5b7f4e487b7100000000000000000000000000000000000000000000000000000000600052604160045260246000fd5b600080fd5b600080fd5b600080fd5b600080fd5b6000601f19601f8301169050919050565b60008160011c9050919050565b7f45524332303a206d696e7420746f20746865207a65726f206164647265737300600082015250565b61125f80620008cc6000396000f3fe608060405234801561001057600080fd5b50600436106100a95760003560e01c80633950935111610071578063395093511461016857806370a082311461019857806395d89b41146101c8578063a457c2d7146101e6578063a9059cbb14610216578063dd62ed3e14610246576100a9565b806306fdde03146100ae578063095ea7b3146100cc57806318160ddd146100fc57806323b872dd1461011a578063313ce5671461014a575b600080fd5b6100b6610276565b6040516100c39190610d20565b60405180910390f35b6100e660048036038101906100e19190610b6a565b610308565b6040516100f39190610d05565b60405180910390f35b61010461032b565b6040516101119190610e22565b60405180910390f35b610134600480360381019061012f9190610b17565b610335565b6040516101419190610d05565b60405180910390f35b610152610364565b60405161015f9190610e3d565b60405180910390f35b610182600480360381019061017d9190610b6a565b61036d565b60405161018f9190610d05565b60405180910390f35b6101b260048036038101906101ad9190610aaa565b6103a4565b6040516101bf9190610e22565b60405180910390f35b6101d06103ec565b6040516101dd9190610d20565b60405180910390f35b61020060048036038101906101fb9190610b6a565b61047e565b60405161020d9190610d05565b60405180910390f35b610230600480360381019061022b9190610b6a565b6104f5565b60405161023d9190610d05565b60405180910390f35b610260600480360381019061025b9190610ad7565b610518565b60405161026d9190610e22565b60405180910390f35b60606003805461028590610f52565b80601f01602080910402602001604051908101604052809291908181526020018280546102b190610f52565b80156102fe5780601f106102d3576101008083540402835291602001916102fe565b820191906000526020600020905b8154815290600101906020018083116102e157829003601f168201915b5050505050905090565b60008061031361059f565b90506103208185856105a7565b600191505092915050565b6000600254905090565b60008061034061059f565b905061034d858285610772565b6103588585856107fe565b60019150509392505050565b60006012905090565b60008061037861059f565b905061039981858561038a8589610518565b6103949190610e74565b6105a7565b600191505092915050565b60008060008373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020549050919050565b6060600480546103fb90610f52565b80601f016020809104026020016040519081016040528092919081815260200182805461042790610f52565b80156104745780601f1061044957610100808354040283529160200191610474565b820191906000526020600020905b81548152906001019060200180831161045757829003601f168201915b5050505050905090565b60008061048961059f565b905060006104978286610518565b9050838110156104dc576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016104d390610e02565b60405180910390fd5b6104e982868684036105a7565b60019250505092915050565b60008061050061059f565b905061050d8185856107fe565b600191505092915050565b6000600160008473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002060008373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002054905092915050565b600033905090565b600073ffffffffffffffffffffffffffffffffffffffff168373ffffffffffffffffffffffffffffffffffffffff161415610617576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161060e90610de2565b60405180910390fd5b600073ffffffffffffffffffffffffffffffffffffffff168273ffffffffffffffffffffffffffffffffffffffff161415610687576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161067e90610d62565b60405180910390fd5b80600160008573ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002060008473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020819055508173ffffffffffffffffffffffffffffffffffffffff168373ffffffffffffffffffffffffffffffffffffffff167f8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925836040516107659190610e22565b60405180910390a3505050565b600061077e8484610518565b90507fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff81146107f857818110156107ea576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016107e190610d82565b60405180910390fd5b6107f784848484036105a7565b5b50505050565b600073ffffffffffffffffffffffffffffffffffffffff168373ffffffffffffffffffffffffffffffffffffffff16141561086e576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161086590610dc2565b60405180910390fd5b600073ffffffffffffffffffffffffffffffffffffffff168273ffffffffffffffffffffffffffffffffffffffff1614156108de576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016108d590610d42565b60405180910390fd5b6108e9838383610a76565b60008060008573ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019081526020016000205490508181101561096f576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161096690610da2565b60405180910390fd5b8181036000808673ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002081905550816000808573ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020600082825401925050819055508273ffffffffffffffffffffffffffffffffffffffff168473ffffffffffffffffffffffffffffffffffffffff167fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef84604051610a5d9190610e22565b60405180910390a3610a70848484610a7b565b50505050565b505050565b505050565b600081359050610a8f816111fb565b92915050565b600081359050610aa481611212565b92915050565b600060208284031215610ac057610abf610fe2565b5b6000610ace84828501610a80565b91505092915050565b60008060408385031215610aee57610aed610fe2565b5b6000610afc85828601610a80565b9250506020610b0d85828601610a80565b9150509250929050565b600080600060608486031215610b3057610b2f610fe2565b5b6000610b3e86828701610a80565b9350506020610b4f86828701610a80565b9250506040610b6086828701610a95565b9150509250925092565b60008060408385031215610b8157610b80610fe2565b5b6000610b8f85828601610a80565b9250506020610ba085828601610a95565b9150509250929050565b610bb381610edc565b82525050565b6000610bc482610e58565b610bce8185610e63565b9350610bde818560208601610f1f565b610be781610fe7565b840191505092915050565b6000610bff602383610e63565b9150610c0a82610ff8565b604082019050919050565b6000610c22602283610e63565b9150610c2d82611047565b604082019050919050565b6000610c45601d83610e63565b9150610c5082611096565b602082019050919050565b6000610c68602683610e63565b9150610c73826110bf565b604082019050919050565b6000610c8b602583610e63565b9150610c968261110e565b604082019050919050565b6000610cae602483610e63565b9150610cb98261115d565b604082019050919050565b6000610cd1602583610e63565b9150610cdc826111ac565b604082019050919050565b610cf081610f08565b82525050565b610cff81610f12565b82525050565b6000602082019050610d1a6000830184610baa565b92915050565b60006020820190508181036000830152610d3a8184610bb9565b905092915050565b60006020820190508181036000830152610d5b81610bf2565b9050919050565b60006020820190508181036000830152610d7b81610c15565b9050919050565b60006020820190508181036000830152610d9b81610c38565b9050919050565b60006020820190508181036000830152610dbb81610c5b565b9050919050565b60006020820190508181036000830152610ddb81610c7e565b9050919050565b60006020820190508181036000830152610dfb81610ca1565b9050919050565b60006020820190508181036000830152610e1b81610cc4565b9050919050565b6000602082019050610e376000830184610ce7565b92915050565b6000602082019050610e526000830184610cf6565b92915050565b600081519050919050565b600082825260208201905092915050565b6000610e7f82610f08565b9150610e8a83610f08565b9250827fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff03821115610ebf57610ebe610f84565b5b828201905092915050565b6000610ed582610ee8565b9050919050565b60008115159050919050565b600073ffffffffffffffffffffffffffffffffffffffff82169050919050565b6000819050919050565b600060ff82169050919050565b60005b83811015610f3d578082015181840152602081019050610f22565b83811115610f4c576000848401525b50505050565b60006002820490506001821680610f6a57607f821691505b60208210811415610f7e57610f7d610fb3565b5b50919050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052601160045260246000fd5b7f4e487b7100000000000000000000000000000000000000000000000000000000600052602260045260246000fd5b600080fd5b6000601f19601f8301169050919050565b7f45524332303a207472616e7366657220746f20746865207a65726f206164647260008201527f6573730000000000000000000000000000000000000000000000000000000000602082015250565b7f45524332303a20617070726f766520746f20746865207a65726f20616464726560008201527f7373000000000000000000000000000000000000000000000000000000000000602082015250565b7f45524332303a20696e73756666696369656e7420616c6c6f77616e6365000000600082015250565b7f45524332303a207472616e7366657220616d6f756e742065786365656473206260008201527f616c616e63650000000000000000000000000000000000000000000000000000602082015250565b7f45524332303a207472616e736665722066726f6d20746865207a65726f20616460008201527f6472657373000000000000000000000000000000000000000000000000000000602082015250565b7f45524332303a20617070726f76652066726f6d20746865207a65726f2061646460008201527f7265737300000000000000000000000000000000000000000000000000000000602082015250565b7f45524332303a2064656372656173656420616c6c6f77616e63652062656c6f7760008201527f207a65726f000000000000000000000000000000000000000000000000000000602082015250565b61120481610eca565b811461120f57600080fd5b50565b61121b81610f08565b811461122657600080fd5b5056fea2646970667358221220a9aee9d41235ecae0d5cb0b5883095406fb676811b99b9a30c8742c6eae9742464736f6c63430008070033';
+export const ERC20_BYTE_CODE = '60806040523480156200001157600080fd5b506040518060400160405280600b81526020017f476f6f6679476f6f6265720000000000000000000000000000000000000000008152506040518060400160405280600281526020017f47470000000000000000000000000000000000000000000000000000000000008152508160039080519060200190620000969291906200024b565b508060049080519060200190620000af9291906200024b565b505050620000cd3368056bc75e2d63100000620000d360201b60201c565b62000499565b600073ffffffffffffffffffffffffffffffffffffffff168273ffffffffffffffffffffffffffffffffffffffff16141562000146576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016200013d906200034e565b60405180910390fd5b6200015a600083836200024160201b60201c565b80600260008282546200016e91906200039e565b92505081905550806000808473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020600082825401925050819055508173ffffffffffffffffffffffffffffffffffffffff16600073ffffffffffffffffffffffffffffffffffffffff167fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef8360405162000221919062000370565b60405180910390a36200023d600083836200024660201b60201c565b5050565b505050565b505050565b828054620002599062000405565b90600052602060002090601f0160209004810192826200027d5760008555620002c9565b82601f106200029857805160ff1916838001178555620002c9565b82800160010185558215620002c9579182015b82811115620002c8578251825591602001919060010190620002ab565b5b509050620002d89190620002dc565b5090565b5b80821115620002f7576000816000905550600101620002dd565b5090565b60006200030a601f836200038d565b91507f45524332303a206d696e7420746f20746865207a65726f2061646472657373006000830152602082019050919050565b6200034881620003fb565b82525050565b600060208201905081810360008301526200036981620002fb565b9050919050565b60006020820190506200038760008301846200033d565b92915050565b600082825260208201905092915050565b6000620003ab82620003fb565b9150620003b883620003fb565b9250827fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff03821115620003f057620003ef6200043b565b5b828201905092915050565b6000819050919050565b600060028204905060018216806200041e57607f821691505b602082108114156200043557620004346200046a565b5b50919050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052601160045260246000fd5b7f4e487b7100000000000000000000000000000000000000000000000000000000600052602260045260246000fd5b6111f680620004a96000396000f3fe608060405234801561001057600080fd5b50600436106100a95760003560e01c80633950935111610071578063395093511461016857806370a082311461019857806395d89b41146101c8578063a457c2d7146101e6578063a9059cbb14610216578063dd62ed3e14610246576100a9565b806306fdde03146100ae578063095ea7b3146100cc57806318160ddd146100fc57806323b872dd1461011a578063313ce5671461014a575b600080fd5b6100b6610276565b6040516100c39190610ebf565b60405180910390f35b6100e660048036038101906100e19190610b5e565b610308565b6040516100f39190610ea4565b60405180910390f35b61010461032b565b6040516101119190610fc1565b60405180910390f35b610134600480360381019061012f9190610b0f565b610335565b6040516101419190610ea4565b60405180910390f35b610152610364565b60405161015f9190610fdc565b60405180910390f35b610182600480360381019061017d9190610b5e565b61036d565b60405161018f9190610ea4565b60405180910390f35b6101b260048036038101906101ad9190610aaa565b6103a4565b6040516101bf9190610fc1565b60405180910390f35b6101d06103ec565b6040516101dd9190610ebf565b60405180910390f35b61020060048036038101906101fb9190610b5e565b61047e565b60405161020d9190610ea4565b60405180910390f35b610230600480360381019061022b9190610b5e565b6104f5565b60405161023d9190610ea4565b60405180910390f35b610260600480360381019061025b9190610ad3565b610518565b60405161026d9190610fc1565b60405180910390f35b606060038054610285906110f1565b80601f01602080910402602001604051908101604052809291908181526020018280546102b1906110f1565b80156102fe5780601f106102d3576101008083540402835291602001916102fe565b820191906000526020600020905b8154815290600101906020018083116102e157829003601f168201915b5050505050905090565b60008061031361059f565b90506103208185856105a7565b600191505092915050565b6000600254905090565b60008061034061059f565b905061034d858285610772565b6103588585856107fe565b60019150509392505050565b60006012905090565b60008061037861059f565b905061039981858561038a8589610518565b6103949190611013565b6105a7565b600191505092915050565b60008060008373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020549050919050565b6060600480546103fb906110f1565b80601f0160208091040260200160405190810160405280929190818152602001828054610427906110f1565b80156104745780601f1061044957610100808354040283529160200191610474565b820191906000526020600020905b81548152906001019060200180831161045757829003601f168201915b5050505050905090565b60008061048961059f565b905060006104978286610518565b9050838110156104dc576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016104d390610fa1565b60405180910390fd5b6104e982868684036105a7565b60019250505092915050565b60008061050061059f565b905061050d8185856107fe565b600191505092915050565b6000600160008473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002060008373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002054905092915050565b600033905090565b600073ffffffffffffffffffffffffffffffffffffffff168373ffffffffffffffffffffffffffffffffffffffff161415610617576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161060e90610f81565b60405180910390fd5b600073ffffffffffffffffffffffffffffffffffffffff168273ffffffffffffffffffffffffffffffffffffffff161415610687576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161067e90610f01565b60405180910390fd5b80600160008573ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002060008473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020819055508173ffffffffffffffffffffffffffffffffffffffff168373ffffffffffffffffffffffffffffffffffffffff167f8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925836040516107659190610fc1565b60405180910390a3505050565b600061077e8484610518565b90507fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff81146107f857818110156107ea576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016107e190610f21565b60405180910390fd5b6107f784848484036105a7565b5b50505050565b600073ffffffffffffffffffffffffffffffffffffffff168373ffffffffffffffffffffffffffffffffffffffff16141561086e576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161086590610f61565b60405180910390fd5b600073ffffffffffffffffffffffffffffffffffffffff168273ffffffffffffffffffffffffffffffffffffffff1614156108de576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016108d590610ee1565b60405180910390fd5b6108e9838383610a76565b60008060008573ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019081526020016000205490508181101561096f576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161096690610f41565b60405180910390fd5b8181036000808673ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002081905550816000808573ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020600082825401925050819055508273ffffffffffffffffffffffffffffffffffffffff168473ffffffffffffffffffffffffffffffffffffffff167fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef84604051610a5d9190610fc1565b60405180910390a3610a70848484610a7b565b50505050565b505050565b505050565b600081359050610a8f81611192565b92915050565b600081359050610aa4816111a9565b92915050565b600060208284031215610abc57600080fd5b6000610aca84828501610a80565b91505092915050565b60008060408385031215610ae657600080fd5b6000610af485828601610a80565b9250506020610b0585828601610a80565b9150509250929050565b600080600060608486031215610b2457600080fd5b6000610b3286828701610a80565b9350506020610b4386828701610a80565b9250506040610b5486828701610a95565b9150509250925092565b60008060408385031215610b7157600080fd5b6000610b7f85828601610a80565b9250506020610b9085828601610a95565b9150509250929050565b610ba38161107b565b82525050565b6000610bb482610ff7565b610bbe8185611002565b9350610bce8185602086016110be565b610bd781611181565b840191505092915050565b6000610bef602383611002565b91507f45524332303a207472616e7366657220746f20746865207a65726f206164647260008301527f65737300000000000000000000000000000000000000000000000000000000006020830152604082019050919050565b6000610c55602283611002565b91507f45524332303a20617070726f766520746f20746865207a65726f20616464726560008301527f73730000000000000000000000000000000000000000000000000000000000006020830152604082019050919050565b6000610cbb601d83611002565b91507f45524332303a20696e73756666696369656e7420616c6c6f77616e63650000006000830152602082019050919050565b6000610cfb602683611002565b91507f45524332303a207472616e7366657220616d6f756e742065786365656473206260008301527f616c616e636500000000000000000000000000000000000000000000000000006020830152604082019050919050565b6000610d61602583611002565b91507f45524332303a207472616e736665722066726f6d20746865207a65726f20616460008301527f64726573730000000000000000000000000000000000000000000000000000006020830152604082019050919050565b6000610dc7602483611002565b91507f45524332303a20617070726f76652066726f6d20746865207a65726f2061646460008301527f72657373000000000000000000000000000000000000000000000000000000006020830152604082019050919050565b6000610e2d602583611002565b91507f45524332303a2064656372656173656420616c6c6f77616e63652062656c6f7760008301527f207a65726f0000000000000000000000000000000000000000000000000000006020830152604082019050919050565b610e8f816110a7565b82525050565b610e9e816110b1565b82525050565b6000602082019050610eb96000830184610b9a565b92915050565b60006020820190508181036000830152610ed98184610ba9565b905092915050565b60006020820190508181036000830152610efa81610be2565b9050919050565b60006020820190508181036000830152610f1a81610c48565b9050919050565b60006020820190508181036000830152610f3a81610cae565b9050919050565b60006020820190508181036000830152610f5a81610cee565b9050919050565b60006020820190508181036000830152610f7a81610d54565b9050919050565b60006020820190508181036000830152610f9a81610dba565b9050919050565b60006020820190508181036000830152610fba81610e20565b9050919050565b6000602082019050610fd66000830184610e86565b92915050565b6000602082019050610ff16000830184610e95565b92915050565b600081519050919050565b600082825260208201905092915050565b600061101e826110a7565b9150611029836110a7565b9250827fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff0382111561105e5761105d611123565b5b828201905092915050565b600061107482611087565b9050919050565b60008115159050919050565b600073ffffffffffffffffffffffffffffffffffffffff82169050919050565b6000819050919050565b600060ff82169050919050565b60005b838110156110dc5780820151818401526020810190506110c1565b838111156110eb576000848401525b50505050565b6000600282049050600182168061110957607f821691505b6020821081141561111d5761111c611152565b5b50919050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052601160045260246000fd5b7f4e487b7100000000000000000000000000000000000000000000000000000000600052602260045260246000fd5b6000601f19601f8301169050919050565b61119b81611069565b81146111a657600080fd5b50565b6111b2816110a7565b81146111bd57600080fd5b5056fea2646970667358221220ebccbd10137697c90dfc487770458b188f7a7a55af48e17d68a6a9bd2e13964864736f6c63430008000033';
