@@ -45,7 +45,7 @@ describeDevNode('precompile_relay_manager - precompile view functions', (context
     );
     const decoded_is_relayer = context.web3.eth.abi.decodeParameters(
       ['bool'],
-      is_relayer.result,
+      is_relayer,
     )[0];
     expect(decoded_is_relayer).equal(true);
 
@@ -59,7 +59,7 @@ describeDevNode('precompile_relay_manager - precompile view functions', (context
     );
     const decoded_is_selected_relayer = context.web3.eth.abi.decodeParameters(
       ['bool'],
-      is_selected_relayer.result,
+      is_selected_relayer,
     )[0];
     expect(decoded_is_selected_relayer).equal(true);
 
@@ -75,7 +75,7 @@ describeDevNode('precompile_relay_manager - precompile view functions', (context
     );
     const decoded_is_relayers = context.web3.eth.abi.decodeParameters(
       ['bool'],
-      is_relayers.result,
+      is_relayers,
     )[0];
     expect(decoded_is_relayers).equal(true);
 
@@ -92,7 +92,7 @@ describeDevNode('precompile_relay_manager - precompile view functions', (context
     );
     const decoded_is_selected_relayers = context.web3.eth.abi.decodeParameters(
       ['bool'],
-      is_selected_relayers.result,
+      is_selected_relayers,
     )[0];
     expect(decoded_is_selected_relayers).equal(true);
 
@@ -109,7 +109,7 @@ describeDevNode('precompile_relay_manager - precompile view functions', (context
     );
     const decoded_is_complete_selected_relayers = context.web3.eth.abi.decodeParameters(
       ['bool'],
-      is_complete_selected_relayers.result,
+      is_complete_selected_relayers,
     )[0];
     expect(decoded_is_complete_selected_relayers).equal(true);
 
@@ -127,7 +127,7 @@ describeDevNode('precompile_relay_manager - precompile view functions', (context
     );
     const decoded_is_previous_selected_relayer = context.web3.eth.abi.decodeParameters(
       ['bool'],
-      is_previous_selected_relayer.result,
+      is_previous_selected_relayer,
     )[0];
     expect(decoded_is_previous_selected_relayer).equal(true);
   });
@@ -141,9 +141,9 @@ describeDevNode('precompile_relay_manager - precompile view functions', (context
       'selected_relayers',
       ['0x1'],
     );
-    const decoded_selected_relayers = context.web3.eth.abi.decodeParameters(
+    const decoded_selected_relayers: any = context.web3.eth.abi.decodeParameters(
       ['address[]'],
-      selected_relayers.result,
+      selected_relayers,
     )[0];
     expect(decoded_selected_relayers.length).equal(1);
     expect(decoded_selected_relayers[0]).equal(alithRelayer.public);
@@ -156,9 +156,9 @@ describeDevNode('precompile_relay_manager - precompile view functions', (context
       'previous_selected_relayers',
       ['0x1', '0x1'],
     );
-    const decoded_previous_selected_relayers = context.web3.eth.abi.decodeParameters(
+    const decoded_previous_selected_relayers: any = context.web3.eth.abi.decodeParameters(
       ['address[]'],
-      previous_selected_relayers.result,
+      previous_selected_relayers,
     )[0];
     expect(decoded_previous_selected_relayers.length).equal(1);
     expect(decoded_previous_selected_relayers[0]).equal(alithRelayer.public);
@@ -171,9 +171,9 @@ describeDevNode('precompile_relay_manager - precompile view functions', (context
       'relayer_pool',
       [],
     );
-    const decoded_relayer_pool = context.web3.eth.abi.decodeParameters(
+    const decoded_relayer_pool: any = context.web3.eth.abi.decodeParameters(
       ['address[]', 'address[]'],
-      relayer_pool.result,
+      relayer_pool,
     );
     expect(decoded_relayer_pool[0][0]).equal(alithRelayer.public);
 
@@ -187,7 +187,7 @@ describeDevNode('precompile_relay_manager - precompile view functions', (context
     );
     const decoded_majority = context.web3.eth.abi.decodeParameters(
       ['uint256'],
-      majority.result,
+      majority,
     )[0];
     expect(Number(decoded_majority)).equal(1);
 
@@ -201,7 +201,7 @@ describeDevNode('precompile_relay_manager - precompile view functions', (context
     );
     const decoded_previous_majority = context.web3.eth.abi.decodeParameters(
       ['uint256'],
-      previous_majority.result,
+      previous_majority,
     )[0];
     expect(Number(decoded_previous_majority)).equal(1);
 
@@ -213,9 +213,9 @@ describeDevNode('precompile_relay_manager - precompile view functions', (context
       'relayer_state',
       [alithRelayer.public],
     );
-    const decoded_relayer_state = context.web3.eth.abi.decodeParameters(
+    const decoded_relayer_state: any = context.web3.eth.abi.decodeParameters(
       ['tuple(address,address,uint256)'],
-      relayer_state.result,
+      relayer_state,
     )[0];
     expect(decoded_relayer_state[0]).equal(alithRelayer.public);
 
@@ -227,9 +227,9 @@ describeDevNode('precompile_relay_manager - precompile view functions', (context
       'relayer_states',
       [],
     );
-    const decoded_relayer_states = context.web3.eth.abi.decodeParameters(
+    const decoded_relayer_states: any = context.web3.eth.abi.decodeParameters(
       ['address[]', 'address[]', 'uint256[]'],
-      relayer_states.result,
+      relayer_states,
     );
     expect(decoded_relayer_states[0][0]).equal(alithRelayer.public);
   });
