@@ -1,6 +1,7 @@
 import { expect } from 'chai';
-import { describeDevNode, INodeContext } from '../set_dev_node';
+
 import { customWeb3Request } from '../providers';
+import { describeDevNode, INodeContext } from '../set_dev_node';
 import { createTransaction } from '../transactions';
 
 export async function getEmptyRawTx(context: INodeContext) {
@@ -14,9 +15,9 @@ export async function getEmptyRawTx(context: INodeContext) {
 describeDevNode('TxPool - Genesis', (context) => {
   it('should be empty', async function () {
     const inspect = await customWeb3Request(context.web3, 'txpool_inspect', []);
-    expect(inspect.result.pending).to.be.empty;
+    expect(inspect.pending).to.be.empty;
     const content = await customWeb3Request(context.web3, 'txpool_content', []);
-    expect(content.result.pending).to.be.empty;
+    expect(content.pending).to.be.empty;
   });
 });
 
@@ -24,19 +25,19 @@ describeDevNode('Txpool - Pending Ethereum transaction', (context) => {
   let txHash: string;
   before('Setup: Create transaction', async () => {
     const rawTx = await getEmptyRawTx(context);
-    txHash = (await customWeb3Request(context.web3, 'eth_sendRawTransaction', [rawTx])).result;
+    txHash = (await customWeb3Request(context.web3, 'eth_sendRawTransaction', [rawTx]));
   });
 
   it('should appear in txpool inspection', async function () {
     const inspect = await customWeb3Request(context.web3, 'txpool_inspect', []);
-    const data = inspect.result.pending['0xf24ff3a9cf04c71dbc94d0b566f7a27b94566cac'][context.web3.utils.toHex(0)];
+    const data = inspect.pending['0xf24ff3a9cf04c71dbc94d0b566f7a27b94566cac'][context.web3.utils.toHex(0)];
     expect(data).to.not.be.undefined;
   });
 
   it('should be marked as pending', async function () {
     const pendingTransaction = (
       await customWeb3Request(context.web3, 'eth_getTransactionByHash', [txHash])
-    ).result;
+    );
     expect(pendingTransaction).to.include({
       blockNumber: null,
       hash: txHash
@@ -45,7 +46,7 @@ describeDevNode('Txpool - Pending Ethereum transaction', (context) => {
 
   it('should appear in txpool content', async function () {
     const content = await customWeb3Request(context.web3, 'txpool_content', []);
-    const data = content.result.pending['0xf24ff3a9cf04c71dbc94d0b566f7a27b94566cac'][context.web3.utils.toHex(0)];
+    const data = content.pending['0xf24ff3a9cf04c71dbc94d0b566f7a27b94566cac'][context.web3.utils.toHex(0)];
     expect(data).to.include({
       blockHash: '0x0000000000000000000000000000000000000000000000000000000000000000',
       blockNumber: null,
@@ -61,14 +62,14 @@ describeDevNode('Txpool - Pending Ethereum transaction', (context) => {
 describeDevNode('Txpool - New block', (context) => {
   before('Setup: Create transaction and empty block', async () => {
     const rawTx = await getEmptyRawTx(context);
-    await context.createBlock({transactions: [rawTx]});
+    await context.createBlock({ transactions: [rawTx] });
     await context.createBlock();
   });
 
   it('should reset the txpool', async function () {
     const inspect = await customWeb3Request(context.web3, 'txpool_inspect', []);
-    expect(inspect.result.pending).to.be.empty;
+    expect(inspect.pending).to.be.empty;
     let content = await customWeb3Request(context.web3, 'txpool_content', []);
-    expect(content.result.pending).to.be.empty;
+    expect(content.pending).to.be.empty;
   });
 });
